+++ R/fetch_citations.R
@@ -20,8 +20,8 @@ fetch_citations_vt <- function(dataset, redownload = TRUE, minimise=FALSE) {
     cli::cli_alert_warning("IDs not necessarily from VecTraits.")
   } else if (!is_from(dataset, "vt")) {
     cli::cli_abort(c(
-      "x" = "IDs not from VecTraits, Please use the {.fn fetch_citations_{get_db(dataset)}} function.",
-      "!" = "Detected db = {.val {get_db(dataset)}}"
+      "x" = "IDs not from VecTraits, Please use the {.fn fetch_citations_{ohvbd_db(dataset)}} function.",
+      "!" = "Detected database = {.val {ohvbd_db(dataset)}}"
     ))
   }
 
@@ -106,8 +106,8 @@ fetch_citations_vd <- function(dataset, redownload = TRUE, minimise = FALSE) {
     cli::cli_alert_warning("IDs not necessarily from VecDyn.")
   } else if (!is_from(dataset, "vd")) {
     cli::cli_abort(c(
-      "x" = "IDs not from VecDyn, Please use the {.fn fetch_citations_{get_db(dataset)}} function.",
-      "!" = "Detected db = {.val {get_db(dataset)}}"
+      "x" = "IDs not from VecDyn, Please use the {.fn fetch_citations_{ohvbd_db(dataset)}} function.",
+      "!" = "Detected database = {.val {ohvbd_db(dataset)}}"
     ))
   }