+++ R/fetch_vd.R
@@ -29,8 +29,8 @@ fetch_vd <- function(ids, rate = 5, connections = 2, basereq = NA) {
     cli::cli_alert_warning("IDs not necessarily from VecDyn.")
   } else if (!is_from(ids, "vd")) {
     cli::cli_abort(c(
-      "x" = "IDs not from VecDyn, Please use the {.fn fetch_{get_db(ids)}} function.",
-      "!" = "Detected db = {.val {get_db(ids)}}"
+      "x" = "IDs not from VecDyn, Please use the {.fn fetch_{ohvbd_db(ids)}} function.",
+      "!" = "Detected database = {.val {ohvbd_db(ids)}}"
     ))
   }