+++ R/fetch_vt.R
@@ -28,8 +28,8 @@ fetch_vt <- function(ids, rate = 5, connections = 2, basereq = NA) {
     cli::cli_alert_warning("IDs not necessarily from VecTraits.")
   } else if (!is_from(ids, "vt")) {
     cli::cli_abort(c(
-      "x" = "IDs not from VecTraits, Please use the {.fn fetch_{get_db(ids)}} function.",
-      "!" = "Detected db = {.val {get_db(ids)}}"
+      "x" = "IDs not from VecTraits, Please use the {.fn fetch_{ohvbd_db(ids)}} function.",
+      "!" = "Detected database = {.val {ohvbd_db(ids)}}"
     ))
   }